Map 1
So far I have textured the map, placed the base down, and placed foliage
down. It is quite a simple map (for the sake of time). And I apoligize
if i went overboard with the foliage again, It is a forest, but I hope
it does not make people computers shit themselves. Here it is